Implement 'remove synonym'.
The 'remove synonym' instruction is a pretty straightforward one to
implement. The only issue is that we're only dealing with oboInOwl-type
synonyms -- because those are the only ones I know about in the
ontologies I work with. But KGCL is not formally tied to OBO (though its
developers are all in the OBO world AFAIK), so this isn't great.