Add Change2TextVisitor.
Add another visitor for Change objects, this one to convert them back
into their textual representation.

This will be used by the ROBOT `apply` command to write back any change
that could not be applied into a "rejected" file.

This may also later be used for a putative `kgcl-diff` command, to print
the differences between two ontologies as a KGCL changeset.