Detect and pass syntax errors to client code.
Add a new simple class to represent a KGCL syntax error. When parsing,
set up custom error listeners to collect any lexical or syntactic errors
that may occur and allow the client code to get the collected list after
parsing.

Change the interface of the KGCLReader's 'read' method to return a
boolean indicating whether parsing was successful or not. The parsed
changeset must now be obtained by a subsequent call to the
'getChangeset' method.

Update the static helper class to reflect the new interface of the
reader and allow collecting the syntax errors.