Add option to disallow partial application of changes.
Add a new `--no-partial-apply` option to the ROBOT command. If set, the
command will only apply changes if *all* the changes from the changeset
can be applied -- if at least one change cannot be applied, the entire
changeset is rejected.

The default behaviour is to apply the changes that can be applied and to
ignore the others.