Allow to use the KGCLReader on strings.
Add a `read(String)` method to KGCLReader to parse KGCL from a string
rather than from a file-like source. For short inputs, this is more
convenient than having to build a StringReader to pass to the
constructor, and it allows to reuse the same KGCLReader object to parse
multiple strings.

When reading from a file-like source, move lexer initialisation to the
constructors. This way, I/O exceptions can only occur when calling the
constructor, not when calling the read() method.