Implement the "move" instruction.
Add support for the NodeMove change. The "node" instruction is now
parsed and applied.

The NodeDeepening and NodeShallowing changes are also implemented. They
are treated as if they were NodeMove changes. We do *not* check whether
the new parent is a proper descendant (in the case of NodeDeepening) or
a proper ascendent (in the case of NodeShallowing) of the existing
parent before proceeding with the change. This may come later.