Add ancestry checking for "deepen" and "shallow" instructions.
Finalise the implementation of the "deepen" and "shallow" instructions
by checking that the new parent is a descendant (deepen) or an ancestor
(shallow) of the existing parent.  This requires the use of a reasoner,
so the ROBOT "apply" command gets a new -r, --reasoner option to specify
the reasoner to use.