Lay the groundwork for using alternative OWL translators.
This commit prepares the code base to allow for the possibility of
having more than one way to render KGCL changes into OWL axioms in the
future.

It introduces a OWLTranslator base class as an implementation of the
IChangeVisitor<List<OWLOntologyChange>> interface. That class itself
does not perform any translation of KGCL into OWL but is intended to act
as a base for classes that will do that. It provides the rejection logic
and may later be expanded to provide more services.

There is currently only one subclass to OWLTranslator,
DirectOWLTranslator, which is basically the old Change2OwlVisitor class.