Remove or rewrite axioms referencing a to-be-obsoleted class.
When obsoleting a class with no replacement (direct or non-direct),
remove all axioms that reference the obsoleted class, instead of leaving
them to be manually fixed by editors. This is apparently what the KGCL
authors expect in that case.

When obsoleting a class with direct replacement, automatically rewrite
all axioms that reference the obsoleted class to make them point to the
replacement class instead. Again, this is seemingly the desired
behaviour.

When obsoleting a class with non-direct replacement, leave the
referencing axioms in place. They need to be kept there so that editors
can find them and manually decide whether they should be rewritten using
one of the suggested non-direct replacements.

closes #2