Avoid rewriting to-be-removed axioms.
Consider the following changeset:

move EX:CHILD from EX:OLD_PARENT to EX:NEW_PARENT
obsolete EX:OLD_PARENT with replacement EX:ANOTHER_CLASS

where we want to replace EX:OLD_PARENT with EX:ANOTHER_CLASS, *except*
for the class EX:CHILD that we want to classify under a completely
different parent. The move operation is performed first, so that
EX:CHILD no longer has any relationship with EX:OLD_PARENT when
EX:OLD_PARENT is obsoleted.

Because KGCL changes are *all* translated to OWLOntologyChange objects
before any of them is actually applied to the ontology, when we
translate the NodeObsoletion the SubCclassOf axiom between EX:CHILD and
EX:OLD_PARENT is still present, and it would therefore be rewritten as
part of the NodeObsoletion procedure, thereby introducing an unwanted
new SubClassOf axiom between EX:CHILD and EX:ANOTHER_CLASS.

To prevent this kind of events, we keep track globally of all the axioms
that are slated for removal in the lifetime of the DirectOWLTranslator,
so that we can avoid rewriting any axiom that is already set to be
removed.