Better handling of non-English labels in obsoletion.
When obsoleting a class, we only add an "obsolete " prefix to the class'
label if the label hs an English language tag or no language tag at all.
Prepending "obsolete " to a label in another language would, in all
likelihood, not make any sense, unless "obsolete" happens to have the
same meaning in that other language as in English.

(*Ideally* we should modify foreign labels in a way that is meaningful
in their respective languages -- for example, prepending "obsolète" to
French labels or "obsoleto"/"obsoleta" to Italian labels. Needless to
say, this is never going to happen.)

If there *is* an English label (or a label without language tag), then
we prepend "obsolete " to it and we remove all the labels in other
languages. Otherwise, if the class has only non-English label(s), then
we preserve them all.