Avoid redundant OWLOntologyChange objects.
When obsoleting a class and removing the axioms that refers to it, take
care of not removing *again* the axioms that have already removed
elsewhere. This does not change the behaviour of the translator (the OWL
ontology manager will simply ignore such redundant changes), but not
having redundant changes will make testing easier.