Allow KGCL objects to be compared on their contents.
We want to be able to compare KGCL objects based on the value of their
fields. That is, if we have for example:

Node n1 = new Node();
n1.setId("N:0001");
Node n2 = new Node();
n2.setId("N:0001");

then we want n1.equals(n2) to be `true`, even if n1 and n2 are different
instances.

For most LinkML-derived classes, this requires to pass the parameter
`callSuper=true` to the Lombok `@EqualsAndHashCode` annotation, so that
the slots from the parent class(es) are taken into account when
comparing instances.

However, if the parent class does not have any slot itself, then the
`callSuper` parameter must be `false`. That's because if the parent
class has no slot, then Lombok will not override the `equals` method, so
what will effectively be called is the `equals` method of Java's
`Object` class -- and that method will return `false` if two objects are
different instances.

So we need to tweak the Java generator script to set the `callSuper`
parameter appropriately depending on whether the parent of the class
currently being generated has slots or not.