Add test fixture for KGCLWriter.
Add a test fixture for the KGCLWriter class. We only test using a
string-backed writer object, so that we do not have to actually write
any file anywhere.

Also slightly rearrange the KGCLReader test to use a callback-based
method as the main helper method, and add a test for the use of a prefix
manager derived from a OWL ontology.