Add "provisional" mode.
We add a mode in which changes are merely recorded in an ontology, as
KGCL annotations, rather than effectively applied. This is to allow
proposed changes to be visible in the ontology for a while (so that
editors can decide whether the changes are good or not) before they are
later applied in a second step.