Record the date of provisional changes.
When working in provisional mode, store the date of the proposed change
(assumed to be the current date, that is the date at which the command
is run) as a dcterms:date annotation.

The --pending option of the apply command now expects a parameter, which
can be either "all" or a date in the YYYY-MM-DD format. If "all", then
all pending changes are extracted from annotations and applied.
Otherwise, only changes with a date older than the specified date are
extracted and applied.