Support parsing/writing type tags.
A literal value may be qualified with a type tag (e.g.
"value"^^xsd:string). We recognise such tags when parsing.

Of note, we do _not_ fill the datatype field with any default value. A
literal that has no type tag is normally assumed to be a xsd:string, but
we do not enforce that at the level of the parser.

This is because downstream applications might need to distinguish
between the case where an explicit type has been specified and the case
where no type has been specified. So the parser cannot enforce the
assumption that "no type tag -> xsd:string" -- that assumption will have
to be done later by consumers of the Change objects.