Lax comparison of language tags (WIP).
As discussed in https://github.com/INCATools/kgcl/issues/60, when trying
to find an annotation value (e.g. to rename a class, we need to find the
annotation corresponding to the old label), language tags should be
compared in a relaxed fashion. We should not fail to find an annotation
just because the annotation has a language tag and the KGCL command did
not specify any language tag at all.

This necessitates some pretty important refactoring, because this means,
among other things, that more than one annotation values may match (if
several annotations have the same literal value but different language
tags, or one has a language tag and another does not).

This is still a work in progress. For now, this is implemented
specifically for the NodeRename operation. After more testing, this will
be generalized to all other operations that involve finding an existing
annotation value (e.g. RemoveDefinition, ChangeDefinition,
SynonymReplacement, etc.).