Implement creation of object and annotation properties and individuals.
Add complete support for the NodeCreation change type and its subtype
ClassCreation and ObjectPropertyCreation.

We add a OwlType enum to represent the corresponding 'owl_type' enum in
the LinkMl model, and we tweak the Jinja template used to generate Java
code so that slots with a 'owl_type' range are rendered as a OwlType
field rather than as a string field (something that should really be
done by the LinkML Java generator, but that's how minimal Java support
is in LinkML).

Update the parser to recognise a single 'create' command with a keyword
indicating which type of node should be created ('class', 'relation' --
aka object property --, 'instance', and 'annotation property').

Update the OWL translator to implement the creation of all types of
nodes.

Of node, the NodeCreation object can store the type of node in two
different places: (1) in its own 'owl_type' slot or (2) in the
'owl_type' slot of the 'about_node' slot. It's unclear why it is so. In
this implementation we only consider the second location.