Change the way prefixes are managed.
When parsing, if a CURIE is using an unknown prefix, do NOT treat it as
if it were an OBO-style prefix (expanding it into
http://purl.obolibrary.org/obo/PREFIX_). Leave the CURIE unexpanded
instead.

OBO-style prefixes are managed at the level of the ROBOT apply command,
where all prefixes known to ROBOT are passed to the KGCL parser. This
gives users the possibility to add their own prefixes using ROBOT's
--prefix option.