Add support for ID range policy files.
Adds a new package for classes dedicated to supporting OBO-style ID
range policy files (*-idranges.owl).

This does not really belong to KGCL-Java, but it's not large enough to
warrant its own library, so having it here will do for now. At least the
classes are in a separate package, so it should be relatively easy to
move them to another project in the future.