ROBOT: Add support for ID range files.
This commit adds a new mode to the kgcl:apply command for automatic
generation of IDs: a mode that is similar to the "manual" mode, except
that the details about the ID range are to be found in a OBO-style ID
range file, instead of being manually specified on the command line.

This mode is selected with the --auto-id-range-file option, which
expects the name of the ID range file. The range name can be specified
with --auto-id-range-name; if no such option is used, the command will
look for a range allocated to "kgcl", "KGCL", "ontobot", or "Ontobot".

If none of the options selecting an ID generation mode
(--auto-id-prefix, --auto-id-temp-prefix, --auto-id-range-file) is used,
the command will by default look for a file whose name ends with
"-idranges.owl". If such a file (only one) is found, the ID range file
mode is automatically enabled, provided a suitable range is found within
the file.