Add a new "mint" command.
Add a new command, initially suggested by Jim Balhoff, to rename
entities that were given temporary identifiers (in a dedicated
namespace) with newly minted permanent identifiers.

This command is not tied to KGCL at all and therefore does not really
belong here, but since (1) it reuses some of the logical already used in
the "apply" command, and (2) it is needed to finalize the work started
by the "apply" command if the --auto-id-temp-prefix option has been
used, it is not *completely* out of place here.

We may think later about moving this command to a separate plugin or
maybe even adding it to the ROBOT core.