Allow using a label where an identifier is expected.
This commit amends the KGCL parser to allow the use of a quoted label
whenever an identifier is expected.

That is, one can now write a KGCL instruction like this:

  obsolete 'my class' with replacement 'another class'

This is purely a feature of the parser. Resolution of labels into
identifiers is performed as part of the parsing, and once parsing is
done, client code will only see Change objects with normal identifiers,
without any way of knowing that the original instruction(s) contained
labels instead.

Resolving the labels into identifiers is delegated to a
IEntityLabelResolver object, that must be provided by the client code to
the parser (by default, the parser has no resolver, so any use of a
label in place of an identifier will lead to an error).