Do not return a change if a parse tree error occured.
If an error occurs when processing the parse tree (which, for now, can
happen for only one possible reason: if a label was used when an
identifier was expected, and the label cannot be resolved into an
identifier), then the change object concerned by the error should not be
returned as if nothing happened.

We thus add a new flag `isBogus` to the parse tree visitor, which is set
to true when an error occurs; before adding the current change to the
list of collected changes, we check that flag and do not add the change
if it is true.