Expose "IDs as labels" feature to client code.
Now that the parser can appropriately deal with the use of labels
whenever identifiers are expected, we can expose that feature to client
code outside of the library.

Helper `parse` methods in KGCLHelper gains a new parameter representing
the label resolver to use when processing the parse tree. Providing that
resolver is the responsability of the caller.

In the ROBOT `apply` command, all we have thus to do is to create a
OntologyBasedLabelResolver object fed with the current ontology, and
pass that object to the KGCLHelper methods.