Add the IPatcher interface.
Add an interface IPatcher based on the public methods of the
OntologyPatcher object. This is to prepare the possibility that we may
add other "patcher" objects in the future, to apply changes to other
types of knowledge graphs beyond OWL ontologies (though there are
currently _no_ plans for doing so).