Remove spurious throws declaration.
The only situation in which the KGCLHelper.parse(String,...) methods can
throw a IOException is if their first argument is null. Instead of
letting that happen, we explicitly test the argument for null-ness, and
have the methods return an empty changeset if the string is indeed null.
We can then safely catch and ignore any IOException that theoretically
could be thrown by the underlying parser, since we know it cannot, in
fact, happen. This allows us to dispense with the "throws IOException"
on those methods.