Do not reset the list of parsed changes when reading from strings.
When using the KGCLReader.read(String) method to parse KGCL instructions
from strings, do not systematically reset the current list of parsed
changes and errors upon each call. Let the changes (and errors)
accumulate across call.

That is, it is possible to do this (as documented, but was previously
not working as intended):

  reader.read("obsolete EX:0001");
  reader.read("create exact synonym 'my synonym' for EX:0002
  List<Change> changeset = reader.getChangeSet();

and the returned changeset object will contain the changes from both
read() calls.