Do not annotate delombed files as generated.
Since we "delombok" the model classes at build time, the only reason we
need Lombok as a runtime dependency is because the "delomboked" classes
are annotated with the "@lombok.generated" annotation.

That annotation may be useful for code analysis, but a nearly 2MB
dependency just for one annotation is certainly not worth it.

So we configure Lombok _not_ to add that annotation, and remove the
entire Lombok runtime dependency. We only need Lombok at build time.