Resolve quoted IDs against OBO shorthands.
When a quoted string is used in place of an identifier, and we need to
lookup the corresponding entity in an ontology, search for matching "OBO
shorthands" in addition to matching labels.

This allows using the OBO shorthand instead of the label (e.g. 'part_of'
instead of 'part of' instead of BFO:0000050), which some users might be
inclined to do given the prevalence of OBO shorthands in some
ontologies (especially ontologies maintained in OBO format).

(Not that I think that using such a shorthand is a good idea.)