Allow 'is_a' as a shorthand for rdfs:subClassOf.
Whenever an ID is expected, we accept a quoted 'is_a' as a shorthand for
rdfs:subClassOf.

In OWL terms, rdfs:subcClassOf is a special relation that is not
comparable to BFO:0000050 and the like, but from a graph point of view,
'is_a' is merely a relation like any other, so since we accept
shorthands like 'part_of' for (non-subsumption) relations, it is
consistent to also accept 'is_a' for subsumption relations.