Consistently use 'about_edge' slot.
The KGCL model is somewhat inconsistent in that the EdgeChange class has
a 'about_edge' slot intended to represent the edge (as a {subject,
predicate, object} triple) that is being changed, but several of its
subclasses (EdgeCreation, EdgeDeletion, MoveUnder, PlaceUnder) _also_
have distinct 'subject', 'predicate', and 'object' slots, which are
redundant with the fields of the 'about_edge' slot.

I believe these slots are not only useless but also harmful, since it is
unclear how the corresponding classes are supposed to be used (i.e.,
should the subject be set in change.subject or in
change.about_edge.subject?). Since I strongly suspect these slots are
actually a mistake (and got no indication of the contrary when I raised
the issue to the KGCL folks [1]), here we update our code to only deal
with the 'about_edge' slot, and completely ignore the "flattened"
subject/object/predicate slots.

[1] https://github.com/INCATools/kgcl/issues/53